Update alt-lab.js
This addresses an issue that appears when a user changes the name of an item in the Menu and it is different from the innerHTML of the H1. Now it instead compares the href link to the URL of the page. Those links should be more permanent as I've seen menu names change a lot.